backend/pyproject.toml
[project]
name = "trading-research-backend"
version = "0.1.0"
description = "Backend MVP for a trading research and execution platform with a single orchestrator agent."
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.13,<2.0",
    "apscheduler>=3.10,<4.0",
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "pydantic-settings>=2.3,<3.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.27,<1.0",
    "pytest>=8.2,<9.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
